feature: change run name to wandb autogenerated name
Currently, the run names are always the output_dir name if not specified by the user.
This leads to duplicated run names in the wandb dashboard.
Wandb auto generates run names.
We can default to auto generated run names if the run name is the same as the output dir name.